Updated readme
Changed the wording to give developers a broader understanding of the problems we're trying to solve, rather than the specifics of how we intend to solve them. Only once developers know the problems we are trying to solve, and whether these are problems that they suffer from, can they know whether it is worth their time to dive deeper and learn about our technical solutions.